Finalizing Laboratory Exercise 1

+++ src/main/java/com/example/emt_labs_201100/web/BookController.java
@@ -2,10 +2,12 @@ package com.example.emt_labs_201100.web;
 
 import com.example.emt_labs_201100.model.Book;
 import com.example.emt_labs_201100.model.dto.BookDto;
+import com.example.emt_labs_201100.model.enums.Category;
 import com.example.emt_labs_201100.service.BookService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Arrays;
 import java.util.List;
 
 @RestController
@@ -48,6 +50,7 @@ public class BookController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Book> deleteById(@PathVariable Long id) {
+        if(this.bookService.findById(id).isEmpty()) return ResponseEntity.notFound().build();
         this.bookService.deleteById(id);
         if(this.bookService.findById(id).isEmpty()) return ResponseEntity.ok().build();
         return ResponseEntity.badRequest().build();
@@ -61,5 +64,10 @@ public class BookController {
                 .orElseGet(() -> ResponseEntity.badRequest().build());
     }
 
+    @GetMapping("/categories")
+    public List<Category> listAllCategories()
+    {
+        return Arrays.stream(Category.values()).toList();
+    }
 
 }
